Introduce authorization for 'Notes' functionality.

+++ EnglishJourney.Application/Note/Commands/DeleteNote/DeleteNoteCommandHandler.cs
@@ -1,11 +1,13 @@
-﻿using EnglishJourney.Domain.Exceptions;
+﻿using EnglishJourney.Domain.Constants;
+using EnglishJourney.Domain.Exceptions;
 using EnglishJourney.Domain.Interfaces;
 using MediatR;
 using Microsoft.Extensions.Logging;
 
 namespace EnglishJourney.Application.Note.Commands.DeleteNote
 {
-    public class DeleteNoteCommandHandler(INoteRepository repository, ILogger<DeleteNoteCommandHandler> logger)
+    public class DeleteNoteCommandHandler(INoteRepository repository, ILogger<DeleteNoteCommandHandler> logger,
+        IEnglishJourneyAuthorizationService authorizationService)
         : IRequestHandler<DeleteNoteCommand>
     {
         public async Task Handle(DeleteNoteCommand request, CancellationToken cancellationToken)
@@ -14,6 +16,9 @@ namespace EnglishJourney.Application.Note.Commands.DeleteNote
 
             var note = await repository.GetById(request.Id);
             if (note == null) throw new NotFoundException(nameof(Domain.Entities.Note), request.Id.ToString());
+
+            if (!authorizationService.AuthorizeNotes(note, ResourceOperation.Delete))
+                throw new ForbidException();
 
             await repository.Delete(note);
         }
